Update footer with MaterialsCloud logo
Serve example notebook without using a separate shell entrypoint file.
Instead, point to the configuration file in `example-notebooks` folder,
which is available in the Docker image as an attached volume.

The Makefile will update the `voila.json` with the chosen template when
invoking either `make serve´ or `make serve_iframe`.